Move multiwallet handling to walletmanager.cpp/.h
