+++ src/wallet/init.h
@@ -7,7 +7,6 @@
 #define BITCOIN_WALLET_INIT_H
 
 #include <string>
-#include <wallet/wallet.h>
 
 class CRPCTable;
 class CScheduler;
@@ -25,42 +24,5 @@ void RegisterWalletRPC(CRPCTable &tableRPC);
 //  This function will perform salvage on the wallet if requested, as long as only one wallet is
 //  being loaded (WalletParameterInteraction forbids -salvagewallet, -zapwallettxes or -upgradewallet with multiwallet).
 bool VerifyWallets();
-
-typedef CWallet* CWalletRef;
-extern std::vector<CWalletRef> vpwallets;
-
-//! Load wallet databases.
-bool OpenWallets();
-
-//! Complete startup of wallets.
-void StartWallets(CScheduler& scheduler);
-
-//! Flush all wallets in preparation for shutdown.
-void FlushWallets();
-
-//! Stop all wallets. Wallets will be flushed first.
-void StopWallets();
-
-//! Close all wallets.
-void CloseWallets();
-
-void AddWallet(CWallet *wallet);
-void DeallocWallet(unsigned int pos);
-
-bool HasWallets();
-
-unsigned int CountWallets();
-
-template<typename Callable>
-void ForEachWallet(Callable&& func)
-{
-    for (auto&& wallet : vpwallets) {
-        func(wallet);
-    }
-};
-
-CWallet* GetWalletAtPos(int pos);
-
-CWallet* FindWalletByName(const std::string &name);
 
 #endif // BITCOIN_WALLET_INIT_H
